pypi release 0.0.1
This is the code contained within the pypi package "bedevere" version 0.0.1. This was my first pypi commit after exploring the testpypi server, and it appears to be working in my own test environment. This is the first true, working commit of the project as the earlier version did not properly import package variables.